This fixes MusicStore #105.
To workaround an EF issue, MusicStore was using name Order (which is a SQL
reserved word) with sql quotes like [Order]. A recent EF check-in added
quotes
around all names which caused it to be double quoted like [[Order]].

Just removing quotes does not fix the issue, because it looks like the EF
check-in
missed quoting some words in some places. So Order is renamed to Orders
for now. @maumar (who investigated this issue today) will do a little more
investigation on EF tomorrow and file a bug on that repo.

This is a workaround to make MusicStore "checkout" scenario work.